+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All versions prior to 0.9.0 are untracked.
 
 ## [Unreleased]
 
+## [1.1.0]
+
 ### Added
 
 * `sigstore sign` now supports Sigstore bundles, which encapsulate the same
@@ -134,7 +136,8 @@ All versions prior to 0.9.0 are untracked.
   ([#351](https://github.com/sigstore/sigstore-python/pull/351))
 
 <!--Release URLs -->
-[Unreleased]: https://github.com/sigstore/sigstore-python/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/sigstore/sigstore-python/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/sigstore/sigstore-python/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/sigstore/sigstore-python/compare/v0.10.0...v1.0.0
 [0.10.0]: https://github.com/sigstore/sigstore-python/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/sigstore/sigstore-python/compare/v0.8.3...v0.9.0